% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreeHiC.R
\name{FreeHiCJuicer}
\alias{FreeHiCJuicer}
\title{FreeHiC directly from hic file}
\usage{
FreeHiCJuicer(
  file,
  chromosomes = NULL,
  pairs = NULL,
  unit = c("BP", "FRAG"),
  seqDepth = NULL,
  countScale = 1,
  noiseRate = 0,
  neighborZeroRate = 0,
  resolution = 50000L
)
}
\arguments{
\item{file}{Filename can be a local path or remote path. 
The remote path full list can be obtained from \url{http://aidenlab.org/data.html}.}

\item{chromosomes}{A vector contains all the chromosomes. 
For example c('chr1', 'chr2'), the resulting contact matrixs will include all the pairs of interaction (chr1_chr1, chr1_chr2, chr2_chr2).}

\item{pairs}{A vector contains all the pair. 
The pair take format as '1_1' or 'chr1_chr1', both means the contact between chromosome1 and chromosome1. If \code{pair} presents, \code{chromosomes} argument will be ignore.}

\item{unit}{Unit only supports c('BP', 'FRAG'). 
'BP' means base-pair, and 'FRAG' means fragment.}

\item{seqDepth}{The desired sequence depth.}

\item{countScale}{The scale of counts. A number larger than 0. If both \code{seqDepth} and \code{countScale} are provided. Choose the larger one.}

\item{noiseRate}{The noise rate for contact matrix. 0 - 1 scale}

\item{neighborZeroRate}{The rate for neighborhood noise rate. 0 - 1 scale}

\item{resolution}{The resolution used in the contacts matrix. A positive number.
The resolution must be a value from following list.
\itemize{
\item{\strong{unit: BP}}
\itemize{
\item 2500000
\item 1000000
\item 500000
\item 250000
\item 100000
\item 50000
\item 25000
\item 10000
\item 5000
}
\item{\strong{unit: FRAG}}
\itemize{
\item 500
\item 250
\item 100
\item 50
\item 20
\item 5
\item 2
\item 1
}
}}
}
\value{
A list contains simulated contact matrix and basic information of hic data
}
\description{
FreeHiC directly from hic file
}
\examples{

library(FreeHiCLite)

## Local file location
localFilePath = system.file('extdata', 'example.hic', package = 'FreeHiCLite')

## Chromosomes needs to be extracted
chromosomes = c('chr1', 'chr2')

## Pairs needs to be extracted
pairs = c('1_1', '1_2')
unit = 'BP'
resolution = 500000L

seqDepth <- 20000
countScale <- 0
noiseRate <- 0.1
neighborZeroRate <- 0


## pass chrosomes into function, it will contains all the interaction pairs

res <- FreeHiCJuicer(file=localFilePath, chromosomes=chromosomes, 
pairs = NULL, unit=unit, resolution=resolution,
seqDepth = seqDepth, countScale = countScale, 
noiseRate = noiseRate, neighborZeroRate = neighborZeroRate)

str(res)


}
