% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juiceReader.R
\name{readJuicer}
\alias{readJuicer}
\title{Juicebox hic file reader}
\usage{
readJuicer(
  file,
  chromosomes = NULL,
  pairs = NULL,
  unit = c("BP", "FRAG"),
  resolution
)
}
\arguments{
\item{file}{Filename can be a local path or remote path. The remote path full list can be obtained from \url{http://aidenlab.org/data.html}.}

\item{chromosomes}{A vector contains all the chromosomes. For example c("chr1", "chr2"), the resulting contact matrixs will include all the pairs of interaction (chr1_chr1, chr1_chr2, chr2_chr2).}

\item{unit}{Unit only supports c("BP", "FRAG"). "BP" means base-pair, and "FRAG" means fragment.}

\item{resolution}{The desired resolution of the contact matrix. The resolution must be a value from following list.
\itemize{
\item{\strong{unit: BP}}
\itemize{
\item 2500000
\item 1000000
\item 500000
\item 250000
\item 100000
\item 50000
\item 25000
\item 10000
\item 5000
}
\item{\strong{unit: FRAG}}
\itemize{
\item 500
\item 250
\item 100
\item 50
\item 20
\item 5
\item 2
\item 1
}
}}

\item{pair}{A vector contains all the pair. The pair take format as "1_1" or "chr1_chr1", both means the contact between chromosome1 and chromosome1. If \code{pair} presents, \code{chromosomes} argument will be ignore.}
}
\value{
A list object includes following items.
\item{contact}{A list contains all the contact matrix. The keys of list is coded as chrA_chrB.}
\item{information}{A list contains basic information of the contact matrix.
\itemize{
\item{genomId} The genom id of the current hic file
\item{resolution} The list of current hic file available resolution.
}}
\item{settings}{A list contains file name, unit, resolution and chromosomes.}
}
\description{
Read \href{https://github.com/aidenlab/juicer/wiki/Data}{.hic} file generated by \href{https://github.com/aidenlab/juicebox}{Juicebox}. Currently can use both local and remote file.
}
\details{
This function is heavily adpoted from both java version juicebox \href{https://github.com/aidenlab/juicer/wiki/Data-Extraction}{Dump} function 
and c++ version \href{https://github.com/aidenlab/straw}{straw}.
}
\examples{

library(FreeHiCLite)

## Remote file location. The reomte file include downloading, it may take a while
remoteFilePath = "https://hicfiles.s3.amazonaws.com/hiseq/gm12878/in-situ/combined.hic"

## Local file location
localFilePath = system.file("extdata", "example.hic", package = "FreeHiCLite")

## Chromosomes needs to be extra
chromosomes = c("chr1", "chr2")

## Pairs needs to be extra
pairs = c("1_1", "1_2")
unit = "BP"
resolution = 500000L

## pass chrosomes into function, it will contains all the interaction pairs

dat <- readJuicer(file=localFilePath, chromosomes=chromosomes, pairs = NULL, unit=unit, resolution=resolution)

print(names(dat[['contact']]))

## pass pairs into function, it will contains only the given pairs
start = Sys.time()
#dat <- readJuicer(file=localFilePath, chromosomes=chromosomes, pairs = pairs, unit=unit, resolution=resolution) ## Give you same result
dat <- readJuicer(file=localFilePath, chromosomes=NULL, pairs = pairs, unit=unit, resolution=resolution)
end = Sys.time()

print(end - start)
str(dat) 

## Access each contact matrix

contacts = dat[['contact']]

## chromsome 1 vs chromosome 1
key = "1_1"
contactMatrix <- contacts[[key]]
head(contactMatrix)

## chromsome 1 vs chromosome 2
key = "1_2"
contactMatrix <- contacts[[key]]
head(contactMatrix)



}
