% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juiceReader.R
\name{readJuicerFragmentSites}
\alias{readJuicerFragmentSites}
\title{Read fragment sites from the given hic file.}
\usage{
readJuicerFragmentSites(file, chromosomes, verbose = TRUE)
}
\arguments{
\item{file}{Filename can be a local path or remote path. The remote path full list can be obtained from \url{http://aidenlab.org/data.html}.}

\item{chromosomes}{A vector contains all the chromosomes. 
For example c('chr1', 'chr2')}

\item{verbose}{A bool value, indicate whether print message}
}
\value{
A list contains all the chromosomes and their fragement sites.
}
\description{
Read \href{https://github.com/aidenlab/juicer/wiki/Data}{.hic} file generated by \href{https://github.com/aidenlab/juicebox}{Juicebox}. Currently can be used for both local and remote file.
}
\details{
This function is heavily adpoted from both java version juicebox \href{https://github.com/aidenlab/juicer/wiki/Data-Extraction}{Dump} function 
and c++ version \href{https://github.com/aidenlab/straw}{straw}.

The \href{https://github.com/aidenlab/juicer/wiki/Pre#restriction-site-file-format}{restricition site file} 
is required for juicer to provide a hic file with fragment level resolution.
}
\examples{
library(FreeHiCLite)

## Remote file location. The reomte file include downloading, it may take a while
remoteFilePath = 'https://hicfiles.s3.amazonaws.com/hiseq/gm12878/in-situ/combined.hic'

## Local file location
localFilePath = system.file('extdata', 'example.hic', package = 'FreeHiCLite')

fragSites <- readJuicerFragmentSites(remoteFilePath, c('chr1', 'chr2'))
str(fragSites)

}
