% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juiceReader.R
\name{readJuicerInformation}
\alias{readJuicerInformation}
\title{Juicebox hic file information}
\usage{
readJuicerInformation(file, verbose = FALSE)
}
\arguments{
\item{file}{Filename can be a local path or remote path. The remote path full list can be obtained from \url{http://aidenlab.org/data.html}.}

\item{verbose}{TRUE or FALSE. Whether print information.}
}
\value{
A list object includes hic file information.
\item{\code{genomId}}{The genome id of the current hic file.}
\item{\code{resolution}}{The list of current hic file available resolution.}
\item{\code{pairs}}{The list of current hic file available pair.}
\item{\code{chromosomeSizes}}{A dataframe of chromosome size (chromosome, size). Can be used in juicer pre function for different genome.}
}
\description{
Read \href{https://github.com/aidenlab/juicer/wiki/Data}{.hic} file generated by \href{https://github.com/aidenlab/juicebox}{Juicebox}. Currently can be used for both local and remote file.
}
\details{
This function is heavily adopted from both java version juicebox \href{https://github.com/aidenlab/juicer/wiki/Data-Extraction}{Dump} function 
and c++ version \href{https://github.com/aidenlab/straw}{straw}.

In case people needs to use juice box \href{https://github.com/aidenlab/juicer/wiki/Pre}{Pre} to generate hic file. 
Check \code{\link[FreeHiCLite]{writeJuicer}} for details.
}
\examples{
library(FreeHiCLite)

## Remote file location. The reomte file include downloading, it may take a while
remoteFilePath = 'https://hicfiles.s3.amazonaws.com/hiseq/gm12878/in-situ/combined.hic'

## Local file location
localFilePath = system.file('extdata', 'example.hic', package = 'FreeHiCLite')

juicerInfo <- readJuicerInformation(localFilePath)
print(str(juicerInfo))

\donttest{
juicerInfo <- readJuicerInformation(remoteFilePath)
print(str(juicerInfo))
}

}
